chore(reorg): Move months creator modal to corresponding Svelte folder
